feat(transformer): support targets option of preset-env
