+++ crates/oxc_transformer/src/es2015/options.rs
@@ -8,3 +8,11 @@ pub struct ES2015Options {
     #[serde(skip)]
     pub arrow_function: Option<ArrowFunctionsOptions>,
 }
+
+impl ES2015Options {
+    #[must_use]
+    pub fn with_arrow_function(mut self, arrow_function: Option<ArrowFunctionsOptions>) -> Self {
+        self.arrow_function = arrow_function;
+        self
+    }
+}
