+++ crates/oxc_transformer/src/lib.rs
@@ -13,6 +13,7 @@ mod compiler_assumptions;
 mod context;
 mod options;
 // Presets: <https://babel.dev/docs/presets>
+mod env;
 mod es2015;
 mod react;
 mod typescript;
@@ -31,8 +32,9 @@ use oxc_span::SourceType;
 use oxc_traverse::{traverse_mut, Traverse, TraverseCtx};
 
 pub use crate::{
-    compiler_assumptions::CompilerAssumptions, es2015::ES2015Options, options::BabelOptions,
-    options::TransformOptions, react::ReactOptions, typescript::TypeScriptOptions,
+    compiler_assumptions::CompilerAssumptions, env::EnvOptions, es2015::ES2015Options,
+    options::BabelOptions, options::TransformOptions, react::ReactOptions,
+    typescript::TypeScriptOptions,
 };
 
 use crate::{
