+++ tasks/transform_conformance/babel_exec.snap.md
@@ -1,10 +1,16 @@
 commit: 4bd1b2c2
 
-Passed: 1/3
+Passed: 8/14
 
 # All Passed:
 * babel-plugin-transform-arrow-functions
 
+
+# babel-preset-env (7/11)
+* plugins-integration/class-arrow-super-tagged-expr/exec.js
+* plugins-integration/issue-15170/exec.js
+* sanity/check-es2015-constants/exec.js
+* sanity/regex-dot-all/exec.js
 
 # babel-plugin-transform-react-jsx-source (0/2)
 * react-source/basic-sample/exec.js
