+++ tasks/transform_conformance/src/constants.rs
@@ -1,4 +1,5 @@
 pub(crate) const PLUGINS: &[&str] = &[
+    "babel-preset-env",
     // // ES2024
     // "babel-plugin-transform-unicode-sets-regex",
     // // ES2022
